+++ NEWS.md
@@ -5,7 +5,8 @@ increments reflect backwards-incompatible API changes, not necessarily big chang
 # huxtable 4.0.0.9000
 
 * Documentation bugfixes
-* Bugfix: `to_html`
+* Bugfix: `to_html` printing left-border: NA in cell CSS
+* Bugfix: `set_all_*` not working when huxtable is not attached
 
 # huxtable 4.0.0