+++ NEWS.md
@@ -5,8 +5,10 @@ increments reflect backwards-incompatible API changes, not necessarily big chang
 # huxtable 4.0.0.9000
 
 * Documentation bugfixes
+* Improved formatting in Excel output
 * Bugfix: `to_html` printing left-border: NA in cell CSS
 * Bugfix: `set_all_*` not working when huxtable is not attached
+
 
 # huxtable 4.0.0